+++ .github/CODEOWNERS
@@ -6,7 +6,7 @@
 
 # review when someone opens a pull request.
 
-- @Fosol @wkaspryk
+- @Fosol
 
 # Order is important; the last matching pattern takes the most
