+++ src/dashboard/package.json
@@ -14,10 +14,10 @@
   },
   "dependencies": {
     "@react-icons/all-files": "^4.1.0",
-    "@types/lodash": "^4.14.202",
+    "@types/lodash": "^4.17.4",
     "@uidotdev/usehooks": "^2.4.1",
-    "axios": "^1.6.2",
-    "chart.js": "^4.4.1",
+    "axios": "^1.7.2",
+    "chart.js": "^4.4.3",
     "classnames": "^2.5.1",
     "cryptr": "^6.3.0",
     "jsonwebtoken": "^9.0.2",
@@ -25,34 +25,34 @@
     "jws": "^4.0.0",
     "jwt-decode": "^4.0.0",
     "lodash": "^4.17.21",
-    "moment": "^2.29.4",
-    "next": "14.0.4",
-    "next-auth": "^4.24.5",
-    "qs": "^6.11.2",
-    "react": "^18.2.0",
+    "moment": "^2.30.1",
+    "next": "14.2.3",
+    "next-auth": "^4.24.7",
+    "qs": "^6.12.1",
+    "react": "^18.3.1",
     "react-chartjs-2": "^5.2.0",
-    "react-dom": "^18.2.0",
-    "react-icons": "^4.12.0",
-    "react-toastify": "^10.0.4",
-    "sharp": "^0.33.0",
-    "zustand": "^4.4.7"
+    "react-dom": "^18.3.1",
+    "react-icons": "^5.2.1",
+    "react-toastify": "^10.0.5",
+    "sharp": "^0.33.4",
+    "zustand": "^4.5.2"
   },
   "devDependencies": {
     "@types/chart.js": "^2.9.41",
-    "@types/jsonwebtoken": "^9.0.5",
+    "@types/jsonwebtoken": "^9.0.6",
     "@types/jwk-to-pem": "^2.0.3",
-    "@types/jws": "^3.2.9",
-    "@types/node": "^20.10.4",
-    "@types/qs": "^6.9.10",
-    "@types/react": "^18.2.43",
-    "@types/react-dom": "^18.2.17",
+    "@types/jws": "^3.2.10",
+    "@types/node": "^20.12.13",
+    "@types/qs": "^6.9.15",
+    "@types/react": "^18.3.3",
+    "@types/react-dom": "^18.3.0",
     "@types/react-icons": "^3.0.0",
-    "autoprefixer": "^10.4.16",
+    "autoprefixer": "^10.4.19",
     "eslint": "^8.55.0",
-    "eslint-config-next": "14.0.4",
-    "postcss": "^8.4.32",
-    "sass": "^1.69.5",
-    "tailwindcss": "^3.3.6",
-    "typescript": "^5.3.3"
+    "eslint-config-next": "14.2.3",
+    "postcss": "^8.4.38",
+    "sass": "^1.77.3",
+    "tailwindcss": "^3.4.3",
+    "typescript": "^5.4.5"
   }
 }
